requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
google-generativeai>=0.3.0
python-dotenv==0.19.0
requests==2.26.0 